pyproject.toml
[project]
name = "postmolodostbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Fedor Spiridonov"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.0"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"